DP_STEEL_PROJECT/requirements.txt
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pdfminer==20191125
Pillow==10.2.0
PyPDF2==3.0.1
tensorflow==2.15.0
tensorflow_intel==2.15.0
tqdm==4.66.1
scipy
pathlib
visualkeras
sqlalchemy==2.0.27
ipykernel
django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
albumentations==1.4.2
easyocr==1.7.1
pandas==2.2.0
uvicorn==0.30.1
whitenoise==6.7.0
PyMuPDF==1.23.15
xlsxwriter==3.2.0